# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"商品管理 API","description":"商品新增、分页查询、详情、修改和删除。价格单位为元。","version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/products/{id}":{"get":{"tags":["商品管理"],"summary":"查询商品详情","operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Product"}}}}}},"put":{"tags":["商品管理"],"summary":"修改商品","description":"完整替换商品名称、价格、库存和描述。","operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Product"}}}}}},"delete":{"tags":["商品管理"],"summary":"删除商品","operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/products":{"get":{"tags":["商品管理"],"summary":"分页查询商品","description":"page 从 0 开始；size 为 1–100；keyword 按名称模糊搜索；按 ID 倒序。","operationId":"list","parameters":[{"name":"keyword","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductPage"}}}}}},"post":{"tags":["商品管理"],"summary":"新增商品","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Product"}}}}}}}},"components":{"schemas":{"ProductRequest":{"type":"object","properties":{"name":{"type":"string","description":"商品名称","example":"无线鼠标","maxLength":100,"minLength":0},"price":{"type":"number","description":"价格，元，最多两位小数","example":59.9,"minimum":0.00},"stock":{"type":"integer","format":"int32","description":"库存数量","example":100,"minimum":0},"description":{"type":"string","description":"商品描述","example":"静音按键，USB 接收器","maxLength":1000,"minLength":0}},"required":["name","price","stock"]},"Product":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"price":{"type":"number"},"stock":{"type":"integer","format":"int32"},"description":{"type":"string"}}},"ProductPage":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/Product"}},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"}}}}}}